v2 Initial Commit
This version renames the entire project from "current monitor" to the more accurate "power monitor" and adds support for a 30 port PCB.  The software now identifies which board is connected to the pi at startup.  As this is a non-passive change, we're moving to version 2.0.0.
The final version of v1 is available in the branch "1.1.3".

Currently the Zero 2 W is the only Pi that has a case ready for the 30 port board.  Other cases to come shortly as well as more information on how to get a 30 port board.  The original 15 port through-hole board will remain the only open source free to use PCB.